+++ core/hoyo/expedition.js
@@ -1,15 +1,13 @@
-import crypto from "crypto";
 import HoyoTemplate from "./template.js";
 
 export default class Expedition extends HoyoTemplate {
-	static DS_SALT = "6s25p5ox5y14umn1p61aqyyvbvvl3lrt";
 	static data = new Map();
 
 	static async checkAndRun (options = {}) {
 		const result = [];
 
 		for (const [uid, account] of Expedition.data) {
-			const generatedDS = Expedition.generateDS(Expedition.DS_SALT);
+			const generatedDS = sr.Utils.generateDS();
 			const region = sr.Utils.getAccountRegion(account.uid);
 
 			const res = await sr.Got({
@@ -122,32 +120,5 @@ export default class Expedition extends HoyoTemplate {
 				fired: false
 			});
 		}
-	}
-
-	static generateDS (salt) {
-		const time = (Date.now() / 1000).toFixed(0);
-		const random = Expedition.randomString();
-		const hash = Expedition.hash(`salt=${salt}&t=${time}&r=${random}`);
-
-		return `${time},${random},${hash}`;
-	}
-
-	static randomString () {
-		let result = "";
-		const chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-		const length = 6;
-
-		for (let i = 0; i < length; i++) {
-			result += chars[Math.floor(Math.random() * chars.length)];
-		}
-
-		return result;
-	}
-
-	static hash (string) {
-		return crypto
-			.createHash("md5")
-			.update(string)
-			.digest("hex");
 	}
 }